# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Life Simulation API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /agents:
    get:
      operationId: listAgents
      tags: [agents]
      summary: Get paginated list of agents
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [name, age, mood, money, currentAction]
        - name: sortDir
          in: query
          schema:
            type: string
            enum: [asc, desc]
        - name: filterAction
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of agents
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentListResponse"
  /agents/{id}:
    get:
      operationId: getAgent
      tags: [agents]
      summary: Get agent profile with needs and relations
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Agent profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentDetail"
        "404":
          description: Agent not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    Agent:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        gender:
          type: string
        age:
          type: integer
        mood:
          type: number
        money:
          type: number
        personality:
          type: string
        socialization:
          type: number
        currentAction:
          type: string
        employerId:
          type: ["integer", "null"]
      required:
        - id
        - name
        - gender
        - age
        - mood
        - money
        - personality
        - socialization
        - currentAction
    AgentNeeds:
      type: object
      properties:
        hunger:
          type: number
        comfort:
          type: number
        social:
          type: number
      required:
        - hunger
        - comfort
        - social
    AgentRelation:
      type: object
      properties:
        otherId:
          type: integer
        otherName:
          type: string
        friendshipLevel:
          type: number
      required:
        - otherId
        - otherName
        - friendshipLevel
    AgentDetail:
      allOf:
        - $ref: "#/components/schemas/Agent"
        - type: object
          properties:
            needs:
              $ref: "#/components/schemas/AgentNeeds"
            relations:
              type: array
              items:
                $ref: "#/components/schemas/AgentRelation"
          required:
            - needs
            - relations
    AgentListResponse:
      type: object
      properties:
        agents:
          type: array
          items:
            $ref: "#/components/schemas/Agent"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required:
        - agents
        - total
        - page
        - limit
        - totalPages
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error