# lib/api-zod/src/generated/types/agent.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Life Simulation API
 * OpenAPI spec version: 0.1.0
 */

export interface Agent {
  id: number;
  name: string;
  gender: string;
  age: number;
  mood: number;
  money: number;
  personality: string;
  socialization: number;
  currentAction: string;
  /** @nullable */
  employerId?: number | null;
}