# lib/api-zod/src/generated/types/agentDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Life Simulation API
 * OpenAPI spec version: 0.1.0
 */
import type { Agent } from "./agent";
import type { AgentNeeds } from "./agentNeeds";
import type { AgentRelation } from "./agentRelation";

export type AgentDetail = Agent & {
  needs: AgentNeeds;
  relations: AgentRelation[];
};