# lib/api-zod/src/generated/types/agentNeeds.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Life Simulation API
 * OpenAPI spec version: 0.1.0
 */

export interface AgentNeeds {
  hunger: number;
  comfort: number;
  social: number;
}